requirements.txt
asn1crypto==1.3.0
astroid==2.3.3
attrs==19.3.0
bitsv==0.11.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
clickclick==1.2.2
coincurve==13.0.0
connexion==2.6.0
dnspython==1.16.0
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-PyMongo==2.3.0
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mnemonic==0.19
openapi-spec-validator==0.2.8
polyglot-bitcoin==0.0.3
pycoin==0.80
pycparser==2.20
pylint==2.4.4
pymongo==3.10.1
pyrsistent==0.16.0
python-bitcoinrpc==1.0
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
swagger-ui-bundle==0.0.6
typed-ast==1.4.1
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
whatsonchain==0.0.3
wrapt==1.11.2
zipp==3.1.0
